+++ packages/ui/components/form/inputs/Input.tsx
@@ -5,8 +5,7 @@ import { useFormContext } from "react-hook-form";
 import classNames from "@calcom/lib/classNames";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 
-import { Alert, Input, InputField, Tooltip } from "../../..";
-import { Eye, EyeOff, Search } from "../../icon";
+import { Alert, Icon, Input, InputField, Tooltip } from "../../..";
 import { Label } from "./Label";
 import type { InputFieldProps } from "./types";
 
@@ -49,9 +48,9 @@ export const PasswordField = forwardRef<HTMLInputElement, InputFieldProps>(funct
             type="button"
             onClick={() => toggleIsPasswordVisible()}>
             {isPasswordVisible ? (
-              <EyeOff className="h-4 stroke-[2.5px]" />
+              <Icon name="eye-off" className="h-4 stroke-[2.5px]" />
             ) : (
-              <Eye className="h-4 stroke-[2.5px]" />
+              <Icon name="eye" className="h-4 stroke-[2.5px]" />
             )}
             <span className="sr-only">{textLabel}</span>
           </button>
@@ -189,7 +188,7 @@ export const FilterSearchField = forwardRef<HTMLInputElement, InputFieldProps>(f
       dir="ltr"
       className="focus-within:ring-brand-default group relative mx-3 mb-1 mt-2.5 flex items-center rounded-md focus-within:outline-none focus-within:ring-2">
       <div className="addon-wrapper border-default [input:hover_+_&]:border-emphasis [input:hover_+_&]:border-l-default [&:has(+_input:hover)]:border-emphasis [&:has(+_input:hover)]:border-r-default flex h-7 items-center justify-center rounded-l-md border border-r-0">
-        <Search className="ms-3 h-4 w-4" data-testid="search-icon" />
+        <Icon name="search" className="ms-3 h-4 w-4" data-testid="search-icon" />
       </div>
       <Input
         ref={ref}